refactor(pack): refactor xdsl deconsolidation
