+++ packages/manager/apps/telecom/src/app/app.less
@@ -104,7 +104,6 @@
 @import './telecom/pack/move/pack-move.less';
 @import './telecom/pack/resiliation/pack-resiliation.less';
 @import './telecom/pack/slots/voipLine/activation/pack-voipLine-activation.less';
-@import './telecom/pack/xdsl/access/deconsolidation/pack-xdsl-access-deconsolidation.less';
 @import './telecom/pack/xdsl/access/linediagnostic/pack-xdsl-access-linediagnostic.less';
 @import './telecom/pack/xdsl/access/pack-xdsl-access.less';
 @import './telecom/pack/xdsl/modem/pack-xdsl-modem.less';
